backend/vision-service/requirements.txt
# Vision Service Dependencies
# All versions are pinned for security and reproducibility

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
Pillow==10.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
cerebras-cloud-sdk==1.23.0
google-generativeai==0.3.0
openai==1.3.0
eth-account==0.9.0
